requirements.txt
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.3
idna==3.11
imageio==2.37.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_loader==0.4
llvmlite==0.45.1
networkx==3.4.2
numba==0.62.1
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
platformdirs==4.5.0
pooch==1.8.2
PyMatting==1.1.14
referencing==0.36.2
rembg==2.0.67
requests==2.32.5
rpds-py==0.27.1
scikit-image==0.25.2
scipy==1.15.3
tifffile==2025.5.10
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
